Add files via upload
In this project we used machine learning to increase sales. We filled this model with an existing data based on previous buyers.  A decision tree is a type of machine learning model that can be used to make decisions based on a set of input variables, which in our case are gender and age. 
When user sign up, they provide their age and gender information. Based on their profile the website will recommend products they will likely to buy
The store is new and due to minimum amount of data available the accuracy is pretty low, but will increase within time.

In this model we separated it into 2 sets: input(age, gender) and output(product type) and tested it. Then we measured our data for accuracy by splitting it into training and testing data and allocated 80% of our data for training and 20% for testing. After we tested it we saved the file and opened it in VS code to create a visual representation of our tree.